Fixing smaller issues in voxel and marching cube scenes

+++ MoravaEngine/SceneProceduralLandmass.h
@@ -81,8 +81,8 @@ private:
 	MapGenerator::MapGenConf m_MapGenConf;
 	MapGenerator::MapGenConf m_MapGenConfPrev;
 
-	float m_HeightMapMultiplier;
-	float m_HeightMapMultiplierPrev;
+	int m_HeightMapMultiplier;
+	int m_HeightMapMultiplierPrev;
 
 	float m_SeaLevel;
 	float m_SeaLevelPrev;
